pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "protocharge"
version = "0.1.0-alpha"
description = "Tools for parsing RESP output and ESP grid data"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "E. Mainas" },
]
dependencies = [
    "numpy>=1.21",
    "networkx>=3.1",
    "rdkit-pypi>=2022.9.5",
    "scipy>=1.12",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "mkdocs>=1.6",
    "pymdown-extensions>=10.0",
]

[project.scripts]
pc = "protocharge.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
